requirements.txt
streamlit
requests
pandas
numpy
matplotlib
python-docx

